requirements.txt
python-dotenv==1.0.0
spotipy==2.23.0
ytmusicapi==1.1.0
